new folder structure created  made some changes in womens ,singleProduct ,cart and other some pages
