+++ frontend/src/component/Card/Card.jsx
@@ -1,19 +1,19 @@
 import React from "react";
 import "./Card.css";
 import { Link, useNavigate } from "react-router-dom";
+import { Box } from "@chakra-ui/react";
 
-export const Card = ({
-  _id,
-  src,
-  brand,
-  title,
-  category,
-  originalPrice,
-  discountPrice,
-  offer,
-}) => {
+export const Card = ({ _id, src, brand, title, category, originalPrice, discountPrice, offer, }) => {
+  const navigate = useNavigate()
+  const handledata = () => {
+    const product = { _id, src, brand, title, category, originalPrice, discountPrice, offer, quantity: 1 }
+    console.log(product)
+    localStorage.setItem("ProductsDetails", JSON.stringify(product))
+    navigate(`/${category}/${_id}`)
+  }
+
   return (
-    <Link to={`/${category}/${_id}`}>
+    <Box onClick={handledata}>
       <div className="card-cont">
         <div className="img-div">
           <img src={src} alt="" />
@@ -27,6 +27,6 @@ export const Card = ({
           </div>
         </div>
       </div>
-    </Link>
+    </Box>
   );
 };
