--- frontend/src/component/Womens/payment.module.css
@@ -1,158 +1,0 @@
-/* 
-        body {
-            margin: 0;
-            padding: 0;
-            font-family: "calibri";
-            font-weight: 400;
-            color: rgb(119, 118, 118);
-            ;
-            line-height: 1.42857143;
-            font-size: 1.4rem;
-            box-sizing: border-box;
-        }
-
-        .span {
-            display: block;
-            font-size: 15px;
-            font-weight: 400;
-        }
-
-        .span::after {
-            display: block;
-            content: " ";
-        }
-
-        h2 {
-            font-weight: 400;
-            text-align: left;
-        }
-
-        #container {
-            margin: 0% 20% 0% 20%;
-
-        }
-
-        button {
-            width: 92%;
-            padding: 20px;
-            background-color: transparent;
-
-            border: solid 1px rgb(51, 51, 51);
-
-        }
-
-        button:hover {
-            color: white;
-            background-color: black;
-        }
-
-        input {
-            width: 300px;
-            padding: 20px;
-            background-color: transparent;
-            border: solid 1px rgb(190, 190, 190);
-
-        }
-
-        #checkout_form {
-            width: 100%;
-        }
-
-        #checkout_form>div {
-            width: 100%;
-            display: inline-Block;
-            margin: 3% 0% 3% 0%;
-        }
-
-        #checkout_form>div>div {
-            display: inline-Block;
-            width: 49%;
-
-        }
-
-        hr {
-            width: 92%;
-            margin: 0;
-            color: rgb(245, 245, 245)
-        }
-
-        .select_title {
-
-            width: 88.5%;
-            padding: 20px;
-            background-color: transparent;
-            border: solid 1px rgb(190, 190, 190);
-
-        }
-
-        .redcolor {
-            color: red;
-
-        }
-
-        #setwidth {
-            width: 88%;
-        }
-
-        h2 {
-            color: #333;
-        }
-
-        #top_nav_bar {
-
-            background-color: rgb(51, 51, 51);
-            color: white;
-            padding: 10px 0px;
-            position: static;
-
-            width: inherit;
-            margin-top: -10px;
-            margin-left: -10;
-            margin-right: -10;
-            top: 0;
-            z-index: 100;
-            font-family: "calibri";
-            font-size: 12px;
-            font-weight: 400;
-            line-height: 12px;
-            text-align: center;
-
-
-
-        }
-
-        #order_details {
-            display: flex;
-            width: 100%;
-        }
-
-        #order_details>div> {
-            width: 300px;
-
-        }
-
-        #order_details>p {
-            width: 300px;
-            font-family: "calibri";
-            font-weight: 400;
-            color: #333;
-            font-size: 15px;
-        }
-
-        .page_header {
-            background-color: #FBF9F9;
-        }
-
-        .page_header>div>a>img {
-            width: 150px;
-        }
-
-        .page_header>div {
-            width: 150px;
-            margin: auto;
-            padding-bottom: 10px;
-            margin-bottom: 30px;
-
-
-        }
-    */
